fix: organizations should delete themselves only

+++ src/organization/organization.controller.ts
@@ -6,10 +6,14 @@ import {
   Post,
   Put,
   Body,
+  UseGuards,
+  Req,
+  ForbiddenException,
 } from '@nestjs/common';
 import { ZodValidationPipe } from 'nestjs-zod';
 import { OrganizationService } from './organization.service';
 import { CreateOrgDto } from './dto/create.organization.dto';
+import { JWTAuthGuard } from 'src/auth/auth.guard';
 
 @Controller('organizations')
 export class OrganizationController {
@@ -33,7 +37,10 @@ export class OrganizationController {
   }
 
   @Delete(':id')
-  removeOrganization(@Param('id') id: string) {
+  @UseGuards(JWTAuthGuard)
+  removeOrganization(@Param('id') id: string, @Req() req: any) {
+    if (id !== req.organization._id)
+      throw new ForbiddenException("You don't have necessary permission.");
     return this.organizationService.deleteOrg(id);
   }
 }
